backend/requirements.txt
django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.9.0
requests==2.33.1
gunicorn==23.0.0
whitenoise==6.9.0
